Refactor administration in own container
